requirements.txt
# Telegram Bot
python-telegram-bot==21.0
python-telegram-bot[job-queue]==21.0

# AI
anthropic==0.18.1

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Cache & Queue
redis==5.0.1
aioredis==2.0.1

# Scheduler
apscheduler==3.10.4

# Web Framework (Admin)
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Payments
yookassa==3.0.1

# Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP
httpx>=0.27.0
aiohttp==3.9.1

# Utils
python-dateutil==2.8.2
pytz==2024.1

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Dev
black==24.1.1
isort==5.13.2
flake8==7.0.0
